Fix @Warning for high precision

+++ Annotations/src/main/java/annotations/tc/TCWarningAnnotation.java
@@ -101,7 +101,7 @@ public class TCWarningAnnotation extends TCAnnotation
 			else
 			{
 				TCIntegerLiteralExpression w = (TCIntegerLiteralExpression)arg;
-				suppressed.add(w.value.value);
+				suppressed.add(w.value.value.longValue());
 			}
 		}
 	}
